Extract all sierra integration tests to files, and merging them.
commit-id:1a7b3d2a
